bumps the comparison script(s) to a separate folder
